refactor: record로 변경 및 패키지 변경 #4
