+++ src/main/java/com/ssafy/foofa/battle/domain/dto/CreateBattleRequest.java
@@ -1,11 +1,10 @@
 package com.ssafy.foofa.battle.domain.dto;
 
 import com.ssafy.foofa.battle.domain.Battle;
-import lombok.Getter;
 
-@Getter
-public class CreateBattleRequest {
-    private String inviteCode;
-    private String hostUserId;
-    private Battle.Settings setting;
+public record CreateBattleRequest(
+        String inviteCode,
+        String hostUserId,
+        Battle.Settings setting
+) {
 }
