+++ src/main/java/com/ssafy/foofa/chat/infra/RedisMessagePublisher.java
@@ -1,6 +1,6 @@
 package com.ssafy.foofa.chat.infra;
 
-import com.ssafy.foofa.chat.presentation.dto.event.ChatMessageEvent;
+import com.ssafy.foofa.chat.application.event.ChatMessageEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
@@ -13,7 +13,7 @@ public class RedisMessagePublisher {
     private static final String CHANNEL_PREFIX = "chat.battle.";
 
     public void publishMessage(ChatMessageEvent event) {
-        String channel = CHANNEL_PREFIX + event.getBattleId();
+        String channel = CHANNEL_PREFIX + event.battleId();
         redisTemplate.convertAndSend(channel, event);
     }
 }
