+++ src/main/java/com/ssafy/foofa/chat/infra/WebSocketMessageSender.java
@@ -1,6 +1,6 @@
 package com.ssafy.foofa.chat.infra;
 
-import com.ssafy.foofa.chat.presentation.dto.event.ChatMessageEvent;
+import com.ssafy.foofa.chat.application.event.ChatMessageEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Component;
